MOBILE-3371 core: Extract paginated source class
