+++ src/core/classes/items-management/routed-items-manager-source.ts
@@ -13,14 +13,12 @@
 // limitations under the License.
 
 import { Params } from '@angular/router';
-import { CoreItemsManagerSource } from './items-manager-source';
+import { CorePaginatedItemsManagerSource } from './paginated-items-manager-source';
 
 /**
  * Routed items collection source data.
  */
-export abstract class CoreRoutedItemsManagerSource<Item = unknown> extends CoreItemsManagerSource<Item> {
-
-    protected hasMoreItems = true;
+export abstract class CoreRoutedItemsManagerSource<Item = unknown> extends CorePaginatedItemsManagerSource<Item> {
 
     /**
      * Get a string to identify instances constructed with the given arguments as being reusable.
@@ -30,102 +28,6 @@ export abstract class CoreRoutedItemsManagerSource<Item = unknown> extends CoreI
      */
     static getSourceId(...args: unknown[]): string {
         return args.map(argument => String(argument)).join('-');
-    }
-
-    /**
-     * Check whether there are more pages to be loaded.
-     *
-     * @returns Whether there are more pages to be loaded.
-     */
-    isCompleted(): boolean {
-        return !this.hasMoreItems;
-    }
-
-    /**
-     * Get the count of pages that have been loaded.
-     *
-     * @returns Pages loaded.
-     */
-    getPagesLoaded(): number {
-        if (this.items === null) {
-            return 0;
-        }
-
-        const pageLength = this.getPageLength();
-        if (pageLength === null) {
-            return 1;
-        }
-
-        return Math.ceil(this.items.length / pageLength);
-    }
-
-    /**
-     * Reset collection data.
-     */
-    reset(): void {
-        this.hasMoreItems = true;
-
-        super.reset();
-    }
-
-    /**
-     * Reload the collection, this resets the data to the first page.
-     */
-    async reload(): Promise<void> {
-        this.dirty = true;
-
-        await this.load();
-    }
-
-    /**
-     * Load more items, if any.
-     */
-    async load(): Promise<void> {
-        if (this.dirty) {
-            const { items, hasMoreItems } = await this.loadPageItems(0);
-
-            this.dirty = false;
-            this.setItems(items, hasMoreItems ?? false);
-
-            return;
-        }
-
-        if (!this.hasMoreItems) {
-            return;
-        }
-
-        const { items, hasMoreItems } = await this.loadPageItems(this.getPagesLoaded());
-
-        this.setItems((this.items ?? []).concat(items), hasMoreItems ?? false);
-    }
-
-    /**
-     * Load page items.
-     *
-     * @param page Page number (starting at 0).
-     * @returns Page items data.
-     */
-    protected abstract loadPageItems(page: number): Promise<{ items: Item[]; hasMoreItems?: boolean }>;
-
-    /**
-     * Get the length of each page in the collection.
-     *
-     * @returns Page length; null for collections that don't support pagination.
-     */
-    protected getPageLength(): number | null {
-        return null;
-    }
-
-    /**
-     * Update the collection items.
-     *
-     * @param items Items.
-     * @param hasMoreItems Whether there are more pages to be loaded.
-     */
-    protected setItems(items: Item[], hasMoreItems = false): void {
-        this.hasMoreItems = hasMoreItems;
-
-        super.setItems(items);
     }
 
     /**
